backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose
pydantic
email-validator